refactor: create a handler for validation a user input
create a handler for validation a user input

MODULE-1-PART-3-gallery
